requirements.txt
psycopg2-binary==2.9.9
boto3==1.34.14
pandas==2.1.4
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy==2.0.25
pydantic==2.5.3
email-validator==2.1.0.post1
